Add ability to watch a generator for changes
This greatly helps when developing a box generator. When the --watch option is
supplied, the generator is re-run with the given args each time the source file
changes. This enables immediate feedback about the changes without having to
manually re-run the script from the command line. Exceptions in the generator
are printed to the user but do not crash the program.