DiscRack: Move calculation into parsing stage to raise proper errors
This allows making the warnings into errors that are visible in the web
frontend; that may be overly strict, but I rather err on the side of
erring out rather than having a user cut a model and curse me after.

The calculation of the h-edge had to be altered slightly to work with
input parameters only.