Fix tables for screen readers
Avoid announcing every row number which is part of the navigation for
proper tables (read spread sheet). This is a stop gap solution as tables
should not be used for layout at all. But this at least make things
better for screen reader users until this is switched to proper CSS
based layout.