Keyboard navigation for collapsible sections
For now there was no way to select and open the collapsed sections in
both the generator groups and the setting groups other as by mouse
click. This break the keyboard only work flow and excludes screen reader
users. Add a tab stop and hook up the collapse/uncollaps function to the
return key.