Announce menu collapse and expand
Role tag added as aria-expanded does not work for clickable headings, this is some what of a work around but will get things accessible for now wihtout requiring too much of a rewrite.  The showHide function now has lines to toggle the aria-expanded tag based on the state of the clickable heading.