Fix boxesserver
commit 0a8dcf0e37880c056d85523cdea97d8c8b166c6e caused the server to *only* respond to
localhost by default.  This causes the Docker to not respond properly.  The old
server had the default server name as '', but the new one set it to 'localhost'.

This fixes that problem.

In addition, This commit changes the docker file so that you can
specify which git repo and commit to check out from.  This allows
trivial testing of other clones.  The default branch is master
and default git repo is Florian's, but you  can now specify
new commits and/or branches to clone with

```docker build --build-arg BUILD_BRANCH=<branch-or-commit> --build-arg BUILD_REPO=<my-repo>```