Drop suppport for ai files
as the bounding box is not correct and other formats like ps and pdf are
very similar and work fine (now).

This should be revered in case ps2ai.ps properly processes the
BoundingBox. Given the fact that it is a PS script itself I don't quite
see how it wants to get access to a comment that's even later in the file. So I don't expect that to happen soon - or at all.

Alternative we could be producing AI files ourselves. It's just a gloryfied
PS file anyway. But that currently not planned.

Resolves: #534