LBRN2Surface: Fix closed path detection
Use the last point of the path instead of the new location after a Move command. Doing the later created wrong line segments trying to close the path when moving back to the start location.

This broke FlexTest2 when exporting in lbrn2 format.

Thanks to Clane Sharke <https://github.com/ClaneSharke>for reporting

Resolves: #695