RoundedBox: Fix --top parameter default
In the --top parameter of the roundedbox generator the value of "none"
is set as the default value, but is not included in the choices list.
So the argparse exists with the error

app.py: error: argument --top: invalid choice: 'none' (choose from 'closed', 'hole', 'lid')
if the default "none" value is used when calling the generator execution.

Co-authored-by: Nikita Akimov <https://github.com/Korchy>