feat: make gridfinity tray layout more user friendly
The code forces outside=True, so the web page is misleading because
it currently uses the Traylayout text for outside=False
which simply says "inner height in mm".

Setting outside=True in the Boxes.buildArgParser would fix this,
but then make an outside toggle available.  Instead of this,
the GridfinityTrayLayout now supplies it's own --h implementation
which also allows for setting Gridfinity height in Gridfinity Units
(which most Gridfinity users will prefer).  The default remains set
to 50mm for backwards compatiblity, but it might make more sense for
the default to be Gridfinity compatible (i.e. 3U).

For Gridfinity, most users will want a stacking lip.  This adds 3.56mm
to the total height. Right now, this is turned off for backwards
compatiblity, but having it be default on would be more intuitive
for most gridfinity users.

Finally, when stacking support is turned on, the inner walls are
automatically reduced if they are in auto mode (i.e. height zero).