feat: multi generator support extended testing capabilities
The current test setup can only verify that generators default output
doesn't change; this change uses a YAML file to automate generation of
examples and supports making custom examples with custom arguments.

This is also reusable to automate creation of multiple boxes.  For
example, you are generating a set of boxes for a silverware drawer
and you can write a single YAML file and then share that file with
others to generate all the boxes.

This allows setting a default panel size (i.e. 300mmx 300mm)
and optimize the cuts across these panels automatically

One challenge is that the webpage allows layouts to be automatically
determined (via JavaScript) but the command-line tool requires the
layout to be fully specified.  This adds a YAML __GENERATE__ tag
that allows the command-line tool to auto generate layouts as well.