fix: photoframe module incorrectly uses logging framework
The photoframe module sets up a logger (this is okay) but then
proceeds to set the loggers level and output handler.  This causes
photoframe to emit logging statements to the console (i.e. visible
when boxes_main is run with --examples) and makes it difficult
for upstream code to turn off undesired logging.