SmallPartsTray2: Fix fingerholes
Length of the fingerholes of the back floor did not match up with the
fingers in the y walls. This went unnoticed for quite a while as the
difference was rather small. But the difference couldlead to a different
number of fingers which made them mismatch a lot.

Thanks to Jim Kuzeja <https://github.com/jkuzeja> for reporting!

Resolves: #626