Added an option for a latched lid to RoundedBox
This adds an additional value for `--top` which is `lid_with_latches`. Using this will add an extra plate that
sits just inside (i.e. underneath) the lid, with four latches on it.
The latches may be moved in and out using four M3 screws.

The lid must be glued together with two smaller plates, to make a single unit. Also, the holes for the screws need to be threaded before they will work - depending on the material it may be possible to self-tap.

I've printed this in 3mm acrylic and it worked fairly well.