Add more colors for cutting order.
I've added Color.CUT, which is an array with (currently) 5 colors in it. The first is Color.INNER_CUT and the
last is Color.OUTER_CUT: the fade between the two is linear.

I've then used this to cut the latches in the correct order,
and I've added (bodged) in an annotation showing the cut
order that's intended.